Revert "Describe VIEW for YDB CLI (#9513)"
This reverts commit 1bfcec46a6012e9c0fcc8726e4cbf027e39d3c11.
