--- ydb/core/grpc_services/service_view.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include <memory>
-
-namespace NKikimr::NGRpcService {
-
-class IRequestOpCtx;
-class IFacilityProvider;
-
-void DoDescribeView(std::unique_ptr<IRequestOpCtx> p, const IFacilityProvider& f);
-
-}
