--- ydb/public/sdk/cpp/client/draft/ut/helpers/ya.make
@@ -1,13 +1,0 @@
-LIBRARY()
-
-PEERDIR(
-    ydb/public/api/grpc
-    ydb/public/api/grpc/draft
-)
-
-SRCS(
-    grpc_services/scripting.cpp
-    grpc_services/view.cpp
-)
-
-END()
