--- ydb/services/view/ya.make
@@ -1,11 +1,0 @@
-LIBRARY()
-
-SRCS(
-    grpc_service.cpp
-)
-
-PEERDIR(
-    ydb/core/grpc_services
-)
-
-END()
