+++ ydb/services/ya.make
@@ -21,7 +21,6 @@ RECURSE(
     rate_limiter
     replication
     tablet
-    view
     ydb
     ymq
 )
