+++ ydb/tests/functional/ydb_cli/canondata/result.json
@@ -227,9 +227,6 @@
     "test_ydb_impex.TestImpex.test_stdin[tsv-additional_args3-row]": {
         "uri": "file://test_ydb_impex.TestImpex.test_stdin_tsv-additional_args3-row_/result.output"
     },
-    "test_ydb_scheme.TestSchemeDescribe.test_describe_view": {
-        "uri": "file://test_ydb_scheme.TestSchemeDescribe.test_describe_view/result.output"
-    },
     "test_ydb_scripting.TestExecuteScriptFromStdinWithWideOutput.test_wide_table": {
         "uri": "file://test_ydb_scripting.TestExecuteScriptFromStdinWithWideOutput.test_wide_table/result.output"
     },
