--- ydb/tests/functional/ydb_cli/canondata/test_ydb_scheme.TestSchemeDescribe.test_describe_view/result.output
@@ -1,4 +1,0 @@
-<view> view
-
-Query text:
-SELECT 42
